+++ src/types/event.d.ts
@@ -1,14 +1,21 @@
 export interface Event {
-  id: number
-  start: number | string
-  end: number | string
-  resourceId: string
-  title: string
+  id: number;
+  start: number | string;
+  end: number | string;
+  resourceId: string;
+  title: string;
   //  TODO: 包含所有的内容信息
-  // content: 
-  resizable: boolean
-  movable: boolean
-  startResizable: boolean
-  endResizable: boolean,
-  type: number
+  // detail:
+  resizable: boolean;
+  movable: boolean;
+  startResizable: boolean;
+  endResizable: boolean;
+  popoverVisible: boolean;
+  conflict: boolean;
+  backgroundColor: string;
 }
+
+// <Schedueler
+//   ...
+//   renderPopover={(event, schulderData, ...) => {}}  // popover content
+//   renderCusomEvent={(event) => return React.Component.} // custom event item